refactor(desktop): extract concern files from App.tsx
Decompose the 1025-line monolithic App.tsx into focused modules:
- desktop-chat-wrappers.tsx: DesktopChatPanel, ConnectedChatSurface
- desktop-panel-map.ts: panelComponents registration map
- desktop-rail-config.ts: RAIL_SECTIONS navigation config
- desktop-command-config.ts: buildDesktopCommands factory
- desktop-routes.tsx: BASE_SIMPLE_MODE_ROUTES with shared stubs
- desktop-settings-route.tsx: SettingsRoute component

Remove inline STUB_* arrays and local useChatApi; both now come from
shared packages (@nous/ui and @nous/transport respectively).
